+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Changes
 
  - CSV Export: Column names in generated CSV files were renamed
+ - TV Show search: If the show's title contains its year, it is remove in the search dialog (#1192)
 
 ### Added